requirements.txt
fastapi 
uvicorn 
python-dotenv
uvicorn[standard]